Implement saving a gig.

+++ GigHub/ViewModel/GigFormViewModel.cs
@@ -1,6 +1,6 @@
 ﻿
-using System.Collections.Generic;
 using GigHub.Models;
+using System.Collections.Generic;
 
 namespace GigHub.ViewModel
 {
@@ -9,7 +9,7 @@ namespace GigHub.ViewModel
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
         public IEnumerable<Genre> Genres { get; set; }
         
     }
